Move get_assert_200 and the like into a new WebTest class
