+++ evap/rewards/tests.py
@@ -1,13 +1,14 @@
 from django.conf import settings
-from django_webtest import WebTest
+from django.core.urlresolvers import reverse
+
+from model_mommy import mommy
+
 from evap.evaluation.models import Course
 from evap.evaluation.models import UserProfile
-from evap.evaluation.tests.test_utils import lastform
+from evap.evaluation.tests.test_utils import WebTest, lastform
 from evap.rewards.models import SemesterActivation
 from evap.rewards.models import RewardPointRedemptionEvent
 from evap.rewards.tools import reward_points_of_user
-from django.core.urlresolvers import reverse
-from model_mommy import mommy
 
 
 class RewardTests(WebTest):
